Finalizes Todo done, adds styling
Styling to divide the information more logically.

TodoDone is now only on a specific element in the page vs the whole
todo. All that's left for todos is to allow a user to edit the text on
them and add validation that all fields have been filled.

Adds beginning of editable todos.