Fixes "ghost" todos and adds more styling
If the catch statement for submitting a new todo is reached, the record
that save created is destroyed, preventing a blank entry from being
added to the view. Also adds required to all fields to prevent users
from entering blank inputs Also adds paper styling to lists for a bit of
visual interest.